Full Integration Of The Leave And Join Message
I added a cache to both the leave as well as the join message, so if multiple people join a guild it wont need to check the data base again. Also added so you could add a custom leave message. Made it all into nice embeds too. Will probably do slight changes for ease of use as well as atheistic ones.